% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit conversion}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{Numeric value or vector in degrees Celsius.}
}
\value{
Numeric value or vector in degrees Fahrenheit.
}
\description{
Convert degrees Celsius to Fahrenheit temperatures.
}
\examples{
temp1 <- C_to_F(10)
temp2 <- C_to_F(c(10, 15, 20))
}
