% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin conversion}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{Numeric value or vector in degrees Celsius.}
}
\value{
Numeric value or vector in Kelvin.
}
\description{
Convert degrees Celsius to Kelvin temperatures.
}
\examples{
temp1 <- C_to_K(0)
temp2 <- C_to_K(c(0, 15, 100))
}
