% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius conversion}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{Numeric value or vector in degrees Fahrenheit.}
}
\value{
Numeric value or vector in degrees Celsius.
}
\description{
Convert degrees Fahrenheit to degrees Celsius temperatures.
}
\examples{
temp1 <- F_to_C(32)
temp2 <- F_to_C(c(32, 68, 72))
}
