% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin conversion}
\usage{
F_to_K(F_temp)
}
\arguments{
\item{F_temp}{Numeric value or vector in degrees Fahrenheit.}
}
\value{
Numeric value or vector in Kelvin.
}
\description{
Convert degrees Fahrenheit to Kelvin temperatures.
}
\examples{
temp1 <- F_to_K(100)
temp2 <- F_to_K(c(100, 150, 200))
}
