% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_C}
\alias{K_to_C}
\title{Kelvin to Celsius conversion}
\usage{
K_to_C(K_temp)
}
\arguments{
\item{K_temp}{Numeric value or vector in Kelvin.}
}
\value{
Numeric value or vector in degrees Celsius.
}
\description{
Convert Kelvin temperatures to degrees Celsius.
}
\examples{
temp1 <- K_to_C(280)
temp2 <- K_to_C(c(280, 290, 300.15))
}
