% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit conversion}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{Numeric value or vector in Kelvin.}
}
\value{
Numeric value or vector in degrees Fahrenheit.
}
\description{
Convert Kelvin temperatures to degrees Fahrenheit.
}
\examples{
temp1 <- K_to_F(280)
temp2 <- K_to_F(c(280, 290, 300.15))
}
